% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_network.R
\name{enrich_network}
\alias{enrich_network}
\title{Test a tidygraph network object for enriched Reactome pathways}
\usage{
enrich_network(network, filter = 0.05, background = NULL)
}
\arguments{
\item{network}{tidygraph object, i.e. one output by \code{build_network} with
Ensembl IDs as node names}

\item{filter}{Desired adjusted p-value threshold for filtering output
pathways. Defaults to 0.05.}

\item{background}{Optional background genes used when testing}
}
\value{
A tibble of significantly enriched Reactome pathways
}
\description{
Test a tidygraph network object for enriched Reactome pathways
}
\references{
See \url{https://www.bioconductor.org/packages/ReactomePA/} for
details on ReactomePA's methods.
}
\seealso{
\url{https://github.com/travis-m-blimkie/networker}
}
