% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ppi_data.R
\name{fetch_ppi_data}
\alias{fetch_ppi_data}
\title{Download and process InnateDB PPI data}
\usage{
fetch_ppi_data(dir = getwd())
}
\arguments{
\item{dir}{Directory where output object should be saved; defaults to current
working directory (from \code{getwd()})}
}
\value{
None; saves output to Rds file.
}
\description{
Pull human PPI data from InnateDB and format for use in building
networks. The data frame of PPIs is saved as an Rds file in the chosen
directory (defaults to the current working directory).
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/networker}
}
