% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{innatedb}
\alias{innatedb}
\title{Human PPI data downloaded from InnateDB}
\format{
A data frame with 323626 rows and 4 columns:
\describe{
\item{ensembl_gene_A}{Ensembl gene ID for the first gene/protein in the
interaction}
\item{hgnc_symbol_A}{HGNC symbol for the first gene/protein in the
interaction}
\item{ensembl_gene_B}{Ensembl gene ID for the second gene/protein in the
interaction}
\item{hgnc_symbol_B}{HGNC symbol for the second gene/protein in the
interaction}
}
}
\usage{
innatedb
}
\description{
A data frame containing human PPI data from InnateDB.
}
\keyword{datasets}
