% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_networks.R
\name{integrate_networks}
\alias{integrate_networks}
\title{Combine two (or more?) gene lists into a single integrated network}
\usage{
integrate_networks(df_a, df_b, col, order, ppi_data = innatedb_exp, seed = 1)
}
\arguments{
\item{df_a}{First data frame of input genes, with the first column containing
Ensembl gene IDs.}

\item{df_b}{Second data frame. The name and contents of its first column
should be the same as \code{df_a}}

\item{col}{Column containing Ensembl gene IDs; must be present in both \code{df_a}
and \code{df_b}}

\item{order}{Desired type of output network ("zero", "first", "min_simple" or
"min_steiner"); passed to \code{build_network()}}

\item{ppi_data}{Data frame of PPI data; must contain rows of interactions as
pairs of Ensembl gene IDs, with columns named "ensembl_gene_A" and
"ensembl_gene_B". Defaults to pre-packaged InnateDB PPI data.}

\item{seed}{Number used in call to \code{set.seed()} to allow for reproducible
network generation}
}
\value{
\code{tidygraph} object for plotting or further analysis
}
\description{
Combine two (or more?) gene lists into a single integrated network
}
\details{
The "min_steiner" method is implemented with the \code{SteinerNet}
package.
}
\references{
See \url{https://cran.r-project.org/web/packages/SteinerNet/index.html}
for details on the Steiner network trimming.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/networker}
}
