% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot an undirected network using ggraph}
\usage{
plot_network(
  network,
  fill_column,
  layout = "kk",
  edge_colour = "grey50",
  edge_alpha = 0.2,
  label = FALSE,
  label_column,
  label_filter = 40,
  label_size = 4,
  label_colour = "black",
  label_face = "bold",
  label_padding = 0.25,
  seed = 1
)
}
\arguments{
\item{network}{\code{tidygraph} object, output from \code{build_network}}

\item{fill_column}{Tidy-select column for mapping node colour. The colour
scale set up is green-white-red, designed to map to fold change values}

\item{layout}{Layout of nodes in the network. Supports all layouts from
\code{ggraph}/\code{igraph}, as well as "force_atlas" (see Details)}

\item{edge_colour}{Edge colour, defaults to "grey50"}

\item{edge_alpha}{Transparency of edges, defaults to 0.2}

\item{label}{Boolean, whether labels should be added to nodes. Defaults to
FALSE. Note when TRUE, only seed nodes receive labels.}

\item{label_column}{Tidy-select column of the network/data to be used in
labeling nodes.}

\item{label_filter}{Degree filter used to determine if a given node should be
labeled. Defaults to 40. This value can be tweaked to reduce the number of
node labels, to prevent the network from being too crowded.}

\item{label_size}{Size of node labels, defaults to 4.}

\item{label_colour}{Colour of node labels, defaults to "black"}

\item{label_face}{Font face for node labels, defaults to "bold"}

\item{label_padding}{Padding around the label, defaults to 0.25 lines.}

\item{seed}{Number used in call to \code{set.seed()} to allow for reproducible
network generation}
}
\value{
An object of class "gg"
}
\description{
Plot an undirected network using ggraph
}
\references{
None.
}
\seealso{
\url{https://www.github.com/travis-m-blimkie/networker}
}
