% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot an undirected PPI network using ggraph}
\usage{
plot_network(
  network,
  fill_column,
  layout = "kk",
  legend = FALSE,
  edge_colour = "grey50",
  edge_alpha = 0.2,
  node_size = c(3, 9),
  int_colour = "grey70",
  label = FALSE,
  label_column,
  label_filter = 0,
  label_size = 5,
  label_colour = "black",
  label_face = "bold",
  label_padding = 0.25,
  min_seg_length = 0.25,
  seed = 1
)
}
\arguments{
\item{network}{\code{tidygraph} object, output from \code{build_network}}

\item{fill_column}{Tidy-select column for mapping node colour. The colour
scale set up is green-white-red, designed to map to fold change values}

\item{layout}{Layout of nodes in the network. Supports all layouts from
\code{ggraph}/\code{igraph}, as well as "force_atlas" (see Details)}

\item{legend}{Should a legend be included? Defaults to FALSE}

\item{edge_colour}{Edge colour, defaults to "grey50"}

\item{edge_alpha}{Transparency of edges, defaults to 0.2}

\item{node_size}{Numeric vector of length two, specifying size range of nodes
(maps to node degree). Default is \code{c(3, 9)}.}

\item{int_colour}{Fill colour for non-seed nodes, i.e. interactors. Defaults
to "grey70"}

\item{label}{Boolean, whether labels should be added to nodes. Defaults to
FALSE. Note when TRUE, only seed nodes receive labels.}

\item{label_column}{Tidy-select column of the network/data to be used in
labeling nodes.}

\item{label_filter}{Degree filter used to determine if a given node should be
labeled. Defaults to 0. This value can be tweaked to reduce the number of
node labels, to prevent the network from being too crowded.}

\item{label_size}{Size of node labels, defaults to 4.}

\item{label_colour}{Colour of node labels, defaults to "black"}

\item{label_face}{Font face for node labels, defaults to "bold"}

\item{label_padding}{Padding around the label, defaults to 0.25 lines.}

\item{min_seg_length}{Minimum length of lines to be drawn from labels to
points. The default specified here is 0.25, half of the normal default
value.}

\item{seed}{Number used in call to \code{set.seed()} to allow for reproducible
network generation. Can be changed to get slightly different layouts.}
}
\value{
An object of class "gg"
}
\description{
Plot an undirected PPI network using ggraph
}
\details{
Any layout supported by ggraph can be specified here - see
\code{?layout_tbl_graph_igraph} for a list of options. Additionally, there is
support for the "force_atlas" method, implemented via the ForceAtlas2
package
}
\references{
See \url{https://github.com/analyxcompany/ForceAtlas2} for details on
this method.
}
\seealso{
\url{https://github.com/travis-m-blimkie/networker}
}
