% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot an undirected PPI network using ggraph}
\usage{
plot_network(
  network,
  fill_column,
  fill_type,
  layout = "kk",
  legend = FALSE,
  fontfamily = "Helvetica",
  edge_colour = "grey50",
  edge_alpha = 0.2,
  node_size = c(3, 9),
  int_colour = "grey70",
  label = FALSE,
  label_column,
  label_filter = 0,
  label_size = 5,
  label_colour = "black",
  hub_colour = "blue2",
  label_face = "bold",
  label_padding = 0.25,
  min_seg_length = 0.25,
  seed = 1,
  ...
)
}
\arguments{
\item{network}{\code{tidygraph} object, output from \code{build_network}}

\item{fill_column}{Tidy-select column for mapping node colour. Designed to
handle continuous numeric mappings (either positive/negative only, or
both), and categorical mappings. Two-sided numeric columns will be mapped
to green and red for negative and positive values, respectively.}

\item{fill_type}{Type of fill mapping to perform for nodes, based on supplied
\code{fill_column}. Options are: "fold_change", "two_sided", "one_sided", or
"categorical".}

\item{layout}{Layout of nodes in the network. Supports all layouts from
\code{ggraph}/\code{igraph}, as well as "force_atlas" (see Details)}

\item{legend}{Should a legend be included? Defaults to FALSE}

\item{fontfamily}{Font to use for labels and legend (if present). Defaults to
"Helvetica".}

\item{edge_colour}{Edge colour, defaults to "grey50"}

\item{edge_alpha}{Transparency of edges, defaults to 0.2}

\item{node_size}{Numeric vector of length two, specifying size range of nodes
(maps to node degree). Default is \code{c(3, 9)}.}

\item{int_colour}{Fill colour for non-seed nodes, i.e. interactors. Defaults
to "grey70"}

\item{label}{Boolean, whether labels should be added to nodes. Defaults to
FALSE. Note when TRUE, only seed nodes receive labels.}

\item{label_column}{Tidy-select column of the network/data to be used in
labeling nodes.}

\item{label_filter}{Degree filter used to determine if a given node should be
labeled. Defaults to 0. This value can be tweaked to reduce the number of
node labels, to prevent the network from being too crowded.}

\item{label_size}{Size of node labels, defaults to 5.}

\item{label_colour}{Colour of node labels, defaults to "black"}

\item{hub_colour}{Colour of node labels for hubs. The top 2\% of nodes (based
on calculated hub score) are highlighted with this colour, if \code{label = TRUE}.}

\item{label_face}{Font face for node labels, defaults to "bold"}

\item{label_padding}{Padding around the label, defaults to 0.25 lines.}

\item{min_seg_length}{Minimum length of lines to be drawn from labels to
points. The default specified here is 0.25, half of the normal default
value.}

\item{seed}{Number used in call to \code{set.seed()} to allow for reproducible
network generation. Can be changed to get slightly different layouts.}

\item{...}{Further parameters to be passed on to \code{ggplot2::theme()}, e.g.
\code{legend.position}}
}
\value{
An object of class "gg"
}
\description{
Plot an undirected PPI network using ggraph
}
\details{
Any layout supported by ggraph can be specified here - see
\code{?layout_tbl_graph_igraph} for a list of options. Additionally, there is
support for the "force_atlas" method, implemented via the ForceAtlas2
package.

The \code{fill_type} argument will determine how the node colour is mapped to
the desired column. "fold_change" represents a special case, where the fill
column is numeric and whose values should be mapped to up (> 0) or down (<
0). "two_sided" and "one_sided" are designed for numeric data that contains
either positive and negative values, or only positive/negative values,
respectively. "categorical" is designed for non-numeric mapping colour.

Node statistics (degree, betweenness, and hub score) and calculated using
the respective functions from the \code{tidygraph} package.
}
\references{
See \url{https://github.com/analyxcompany/ForceAtlas2} for details on
this method.
}
\seealso{
\url{https://github.com/travis-m-blimkie/networker}
}
